# Enable embedding through multiple layers of views recursively