--- a/test/fixtures/privileges.sql
+++ b/test/fixtures/privileges.sql
@@ -53,6 +53,8 @@ GRANT ALL ON TABLE
     , public.public_consumers
     , public.public_orders
     , consumers_view
+    , consumers_view_view
+    , consumers_private_view
     , orders_view
     , images
     , images_base64