--- a/test/fixtures/schema.sql
+++ b/test/fixtures/schema.sql
@@ -178,11 +178,11 @@ create view consumers_view as
 create view consumers_view_view as
   select * from consumers_view;
 
-create view private.consumers_private as
+create view public.consumers_extra as
   select * from consumers_view;
 
-create view consumers_private_view as
-  select * from private.consumers_private;
+create view consumers_extra_view as
+  select * from public.consumers_extra;
 
 --
 -- Name: getitemrange(bigint, bigint); Type: FUNCTION; Schema: test; Owner: -