@wolfgangwalther 2020-10-16
> *well not really. not nearly all of those columns are unique or referenced at all in a fk constraint...

Thinking this further... we don't need to return **all** source columns from this query, but only those that are used in any foreign key constraint on either side. Correct?

I changed both the old and the new query to return only those columns:
- The old query now returns 543 rows
- The new query returns 519 rows (the same logic as above applies here)
- The except query returns a difference of 24 columns

Those 24 columns now really represent new opportunities to embed.

And even better - those queries perform better than ever:

```
query        | plan_avg | execution_avg
master       | 0.54ms   | 387.7ms
recursive    | 0.62ms   | 393.0ms
master_fk    | 1.16ms   | 362.2ms
recursive_fk | 1.06ms   | 362.0ms
```

Now, the recursive query is just as fast as the one without recursion - but both are about 7% faster than before.