@wolfgangwalther 2020-10-17
Reordered the whole query a bit and replaced `regexp_split_to_array` with `string_to_array`, which is a lot faster. The query now comes it at around 220-230ms, so almost twice as fast.

Should be good to merge.