@steve-chavez 2020-10-17
Mindblown! :boom:  :exploding_head:

Amazing work here Wolfgang. I wonder what @monacoremo thinks :smile:.

We might just call this new version **The Performance Release** .

I'm going to give it a review and then merge. It's all looking good though, I'm just going to digest and marvel at how the new queries work :)