@steve-chavez 2020-10-25
@wolfgangwalther I've finally understood your great work here!

The query is looking solid, but I think we can clarify the code a bit(for example `allSourceColumns` is no more, it should be something like `allSourceFkColumns`).

I'll do some reviews, but first..

> Embedding is now possible through hidden views in the middle of the chain as well.

Would this cover a realistic use case? 

IIUC, It would imply that a user creates **public** views for their API and then creates **private** views based on those **public** views.

That wouldn't make sense according to [schema isolation](http://postgrest.org/en/v7.0.0/schema_structure.html). The idea is to work inside out,  selecting **private** tables or views to be exposed as **public**.

> The query, as it is now, has to parse all views in the database to detect relationships through hidden views. In a setup with many hidden, but unrelated views this could increase the time for this query.

The price to pay seems a bit too high. PostgREST can be used on legacy databases and those can have a potentially large number of views.