@wolfgangwalther 2020-10-25
> The query is looking solid, but I think we can clarify the code a bit(for example `allSourceColumns` is no more, it should be something like `allSourceFkColumns`).

I'll change that together with everything else you find in your review :)

> Would this cover a realistic use case?
> 
> IIUC, It would imply that a user creates **public** views for their API and then creates **private** views based on those **public** views.
> 
> That wouldn't make sense according to [schema isolation](http://postgrest.org/en/v7.0.0/schema_structure.html). The idea is to work inside out, selecting **private** tables or views to be exposed as **public**.

I did this, because this was exactly what lead me to this issue. It is the other way around, though. My scenario:
- hidden schema `data` with tables
- hidden schema `hidden` with some views as an abstraction on the data tables (some code that is shared between some of my public views)
- public schema `public` with view that use those in `hidden`

So this is exactly the inside-out approach you're describing.

Before my latest change the view detection would only take those views in the `public` schema, but not those in the `hidden` schema. The table in `data` would never be found.

> The price to pay seems a bit too high. PostgREST can be used on legacy databases and those can have a potentially large number of views.

I think, in fact, it probably only makes sense to do it like this. I would assume that it is more likely to have such a "view-chain" when using something like I did above (using views as another layer between data tables and public views).

But maybe we can do something inbetween: Right now we were either looking for views only in the schemas mentioned in `db-schema` or in all schemas everywhere. We could just look for views that are in the search path, so `db-schema` + `db-extra-search-path`. In this case I could add `hidden` to the search path in my case (which is something that I already did anyway). Users can still control how many views are exposed like this, then.